Working on restructure and comments
